requirements.txt
aiocron==1.4
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
croniter==1.0.12
discord==1.0.1
discord.py==1.7.1
Flask==1.1.2
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
numpy==1.20.2
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3 
